ScmLocator: allow plugging in a fallback ScmLocator
